feat(ProductResponseMapper): ProductResponseMapper 추가 - allergens 필드는 AllergyCategory와 중복이 되어 주석처리 하였습니다.

+++ src/main/java/jeje/work/aeatbe/dto/product/ProductResponseDTO.java
@@ -8,7 +8,7 @@ public record ProductResponseDTO(
         String name,
         Long price,
         String imgUrl,
-        Float rating,
+        Double rating,
         String ProductUrl,
         String description,
         String[] freeFrom,
